interface/requirement.txt
streamlit
numpy
opencv-python-headless
qiskit
qiskit-aer
pillow
scikit-learn
torch
torchvision
requests
tqdm